+++ src/app.d.ts
@@ -8,6 +8,7 @@ declare global {
 				email: string;
 				username: string;
 			};
+			isAdmin?: boolean;
 		}
 
 		interface PageData {